Add a dummy .bat file to make output cleaner
Previously I was doing dummy avr-size commands. The advantage is that used an existing program but it cluttered up the output terribly.

This solution is Windows-only so I will need to find a per-OS solution before publishing.

I tried using other built-in options that work in cmd but for some reason not in the IDE (all give "executable not found in %path%" error):
- rem
- echo
- :