Add a dummy library
The idea is to allow the library's example sketch to be compiled to demonstrate how recipe.S.o.pattern, recipe.c.o.pattern, and recipe.cpp.o.pattern are used during the sketch and library compilation steps.

This is working for the sketch, but the library is not compiled. I suspect this may be caused by my dummy recipe.preproc.macros command.