Add .development flag files to bundled libraries
This file makes the example sketches of the libraries writable by the Arduino IDE, which is necessary to allow Sketch > Export compiled Binary to work with them. When the example is in read-only mode, "Export compiled Binary" is canceled with a vague message:

Export canceled, changes must first be saved.

and it's necessary to save a copy of the example outside the library before you can use "Export compiled Binary" with it.

Since it's required to do an "Export compiled Binary" in order to see how properties are handled in all patterns, and it's recommended in the readme to use the example sketches of the libraries for these processes, it makes sense to allow the user to use them in place, rather than increasing the complexity of the instructions to save the example. Although making the examples read-only is a good idea to prevent beginners from inadvertently making changes to them, that is not a concern with the users of this project.