Add properties_list alias to use in the compilation recipes
Because properties are only visible to tools when prepended with tools.TOOL_NAME, it is necessary to name the properties list property tools.dummy_tool.properties_list. However, I feel that property name is somewhat confusing in the compilation recipes, so I created an alias property named properties_list for use in the compilation recipes.